Move lists keydown logic to the slate-lists package
For easy reusability
