+++ packages/slate-lists/package.json
@@ -55,6 +55,7 @@
     },
     "dependencies": {
         "@prezly/slate-commons": "^0.29.4",
+        "is-hotkey": "^0.2.0",
         "uuid": "^8.3.0"
     },
     "devDependencies": {
