+++ packages/slate-lists/src/index.ts
@@ -1,5 +1,6 @@
 export * as Lists from './lib';
 export * as Normalizations from './normalizations';
+export { onKeyDown } from './onKeyDown';
 export * from './types';
 export { withLists } from './withLists';
 export { withListsReact } from './withListsReact';
